backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["David Saidov <delightxxls@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
gunicorn = "^22.0.0"
djoser = "^2.2.2"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-redis = "^5.4.0"
django-import-export = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
